Move MovingTimeTagCaretMode to caret algorithm namespace.
